requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.4
mysqlclient==2.1.1
python-dotenv==1.0.0
Werkzeug==2.2.3 